CHAD-17067: zigbee-fan lazy load subdrivers
