+++ drivers/SmartThings/zigbee-fan/src/fan-light/can_handle.lua
@@ -1,0 +1,14 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local function can_handle_itm_fanlight(opts, driver, device)
+  local FINGERPRINTS = require("fan-light.fingerprints")
+  for _, fingerprint in ipairs(FINGERPRINTS) do
+    if device:get_manufacturer() == fingerprint.mfr and device:get_model() == fingerprint.model then
+      return true, require("fan-light")
+    end
+  end
+  return false
+end
+
+return can_handle_itm_fanlight
