+++ drivers/SmartThings/zigbee-fan/src/fan-light/fingerprints.lua
@@ -1,0 +1,8 @@
+-- Copyright 2025 SmartThings, Inc.
+-- Licensed under the Apache License, Version 2.0
+
+local FINGERPRINTS = {
+  { mfr = "Samsung Electronics", model = "SAMSUNG-ITM-Z-003" },
+}
+
+return FINGERPRINTS
